# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "API Kebutuhan untuk Portal Teknologi",
        "title": "API Teknologi",
        "contact": {},
        "version": "1.0"
    },
    "host": "localhost:3000",
    "basePath": "/api",
    "paths": {
...
        "/api/getdraft-prod_dailies": {
            "get": {
                "description": "Mengambil data produksi dengan status draft",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Data Produksi"
                ],
                "summary": "Get Draft Data",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Page number",
                        "name": "page",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "Limit per page",
                        "name": "limit",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Tanggal periode (format: YYYY-MM-DD)",
                        "name": "tanggal_prod",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Tanggal awal periode (format: YYYY-MM-DD)",
                        "name": "tanggal_awal",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Tanggal akhir periode (format: YYYY-MM-DD)",
                        "name": "tanggal_akhir",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Kode Material",
                        "name": "kode_material",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "Status",
                        "name": "status",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/models.GlobalSuccessHandlerResp"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/definitions/models.ProdDailies"
                                            }
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/models.GlobalErrorHandlerResp"
                        }
                    }
                }
            }
        },
        "/api/getrilis-prod_dailies": {
            "get": {
                "description": "Mengambil data produksi dengan status rilis",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Data Produksi"
                ],
                "summary": "Get Released Data",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Page number",
                        "name": "page",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "Limit per page",
                        "name": "limit",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Tanggal periode (format: YYYY-MM-DD)",
                        "name": "tanggal_prod",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Tanggal awal periode (format: YYYY-MM-DD)",
                        "name": "tanggal_awal",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Tanggal akhir periode (format: YYYY-MM-DD)",
                        "name": "tanggal_akhir",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Kode Material",
                        "name": "kode_material",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "Status",
                        "name": "status",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/models.GlobalSuccessHandlerResp"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/definitions/models.ProdDailies"
                                            }
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/models.GlobalErrorHandlerResp"
                        }
                    }
                }
            }
        },
        "/api/import-prod_dailies": {
            "post": {
                "description": "Mengimpor data dari file Excel dan menyimpannya ke database",
                "consumes": [
                    "multipart/form-data"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Data Produksi"
                ],
                "summary": "Import Excel Data",
                "parameters": [
                    {
                        "type": "file",
                        "description": "File Excel yang akan diimpor",
                        "name": "file",
                        "in": "formData",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/models.GlobalSuccessHandlerResp"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/models.GlobalErrorHandlerResp"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/models.GlobalErrorHandlerResp"
                        }
                    }
                }
            }
        },
        "/api/revoke-prod_dailies": {
            "put": {
                "description": "Mengupdate status produksi berdasarkan metode tertentu",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Data Produksi"
                ],
                "summary": "Revoke Production Data",
                "parameters": [
                    {
                        "description": "Payload untuk update status",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/models.UpdateProdDailiesRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/models.GlobalSuccessHandlerResp"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/models.GlobalErrorHandlerResp"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/models.GlobalErrorHandlerResp"
                        }
                    }
                }
            }
        },
        "/api/update-prod_dailies": {
            "put": {
                "description": "Mengupdate data produksi berdasarkan metode tertentu",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Data Produksi"
                ],
                "summary": "Update Production Data",
                "parameters": [
                    {
                        "description": "Payload untuk update data",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/models.UpdateProdDailiesRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/models.GlobalSuccessHandlerResp"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/models.GlobalErrorHandlerResp"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/models.GlobalErrorHandlerResp"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "models.GlobalErrorHandlerResp": {
            "type": "object",
            "properties": {
                "message": {
                    "type": "string"
                },
                "status": {
                    "type": "boolean"
                }
            }
        },
        "models.GlobalSuccessHandlerResp": {
            "type": "object",
            "properties": {
                "data": {},
                "limit": {},
                "message": {},
                "page": {},
                "status": {
                    "type": "boolean"
                },
                "total": {}
            }
        },
        "models.Material": {
            "type": "object",
            "properties": {
                "createdAt": {
                    "type": "string"
                },
                "deleted_at": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "kodeMat": {
                    "type": "string"
                },
                "material": {
                    "type": "string"
                },
                "updatedAt": {
                    "type": "string"
                }
            }
        },
        "models.Plant": {
            "type": "object",
            "properties": {
                "createdAt": {
                    "type": "string"
                },
                "deleted_at": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "plant": {
                    "type": "string"
                },
                "plantDesc": {
                    "type": "string"
                },
                "updatedAt": {
                    "type": "string"
                }
            }
        },
        "models.ProdDailies": {
            "type": "object",
            "properties": {
                "Material": {
                    "$ref": "#/definitions/models.Material"
                },
                "Plant": {
                    "description": "Definisi relasi",
                    "allOf": [
                        {
                            "$ref": "#/definitions/models.Plant"
                        }
                    ]
                },
                "createdAt": {
                    "type": "string"
                },
                "created_by": {
                    "type": "string"
                },
                "deleted_at": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "jmlProd": {
                    "type": "number"
                },
                "kodeMaterial": {
                    "type": "string"
                },
                "status": {
                    "type": "integer"
                },
                "tglProd": {
                    "type": "string"
                },
                "updatedAt": {
                    "type": "string"
                },
                "uuid": {
                    "type": "string"
                },
                "wcid": {
                    "type": "string"
                }
            }
        },
        "models.UpdateProdDailiesRequest": {
            "type": "object",
            "properties": {
                "data": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/models.ProdDailies"
                    }
                },
                "method": {
                    "type": "string"
                }
            }
        }
    }
}